refactor(client-runtime): share voice supervisor host
